requirements.txt
hydra-core==1.1.1
matplotlib==3.3.2
numpy==1.22.0
omegaconf==2.1.1
tensorflow==2.7.0
